Refactor common utilities: add string trimming functions, container utilities, and hash utilities
