+++ src/common/util/include/openvino/util/common_util.hpp
@@ -4,20 +4,52 @@
 
 #pragma once
 
-#include <algorithm>
-#include <array>
-#include <cctype>
 #include <charconv>
-#include <cstdint>
-#include <cstring>
-#include <filesystem>
-#include <numeric>
+#include <iterator>
+#include <locale>
 #include <optional>
 #include <sstream>
 #include <string>
+#include <string_view>
+#include <type_traits>
 #include <vector>
 
+#include "openvino/util/container_util.hpp"
+#include "openvino/util/hash_util.hpp"
+#include "openvino/util/math_util.hpp"
+
 namespace ov::util {
+
+/**
+ * @brief Return a string view with leading whitespace removed.
+ *
+ * @param s A string view to trim.
+ * @return A string view with leading whitespace removed.
+ */
+constexpr std::string_view ltrim(std::string_view s) {
+    const auto not_ws_pos = s.find_first_not_of(" \t\n\r\f\v");
+    return not_ws_pos == std::string_view::npos ? std::string_view() : s.substr(not_ws_pos);
+}
+
+/**
+ * @brief Return a string view with trailing whitespace removed.
+ * @param s A string view to trim.
+ * @return A string view with trailing whitespace removed.
+ */
+constexpr std::string_view rtrim(std::string_view s) {
+    const auto not_ws_pos = s.find_last_not_of(" \t\n\r\f\v");
+    return not_ws_pos == std::string_view::npos ? std::string_view() : s.substr(0, not_ws_pos + 1);
+}
+
+/**
+ * @brief Return a string view with leading and trailing whitespace removed.
+ * @param s A string view to trim.
+ * @return A string view with leading and trailing whitespace removed.
+ */
+constexpr std::string_view trim(std::string_view s) {
+    return ltrim(rtrim(s));
+}
+
 /**
  * @brief  Helper struct to join container's elements to string using user string as separator.
  * The output string is generated in operator<< and can be used in any context where std::string is expected.
@@ -84,84 +116,7 @@ auto join(const Container& c, std::string_view sep = ", ") {
  */
 template <typename T, typename A>
 std::string vector_to_string(const std::vector<T, A>& v) {
-    return "[ " + ov::util::join(v) + " ]";
-}
-
-std::string to_lower(std::string_view s);
-
-std::string to_upper(std::string_view s);
-
-inline size_t hash_combine(size_t val, const size_t seed) {
-    return seed ^ (val + 0x9e3779b9 + (seed << 6) + (seed >> 2));
-}
-
-inline size_t hash_combine(const std::vector<size_t>& list) {
-    size_t seed = 0;
-    for (size_t v : list) {
-        seed ^= hash_combine(v, seed);
-    }
-    return seed;
-}
-
-inline size_t hash_combine(std::initializer_list<size_t>&& list) {
-    size_t seed = 0;
-    for (size_t v : list) {
-        seed ^= hash_combine(v, seed);
-    }
-    return seed;
-}
-
-constexpr uint64_t u64_hash_combine(uint64_t h, uint64_t k) {
-    // Hash combine formula from boost for uint64_t.
-    constexpr uint64_t m = 0xc6a4a7935bd1e995;
-    constexpr int r = 47;
-
-    k *= m;
-    k ^= k >> r;
-    k *= m;
-
-    h ^= k;
-    h *= m;
-
-    return h + 0xe6546b64;
-}
-
-constexpr uint64_t u64_hash_combine(uint64_t seed, std::initializer_list<uint64_t>&& values) {
-    uint64_t h = seed;
-    for (uint64_t k : values) {
-        h = u64_hash_combine(h, k);
-    }
-    return h;
-}
-
-/**
- * @brief Return a string view with leading whitespace removed.
- *
- * @param s A string view to trim.
- * @return A string view with leading whitespace removed.
- */
-constexpr std::string_view ltrim(std::string_view s) {
-    const auto not_ws_pos = s.find_first_not_of(" \t\n\r\f\v");
-    return not_ws_pos == std::string_view::npos ? std::string_view() : s.substr(not_ws_pos);
-}
-
-/**
- * @brief Return a string view with trailing whitespace removed.
- * @param s A string view to trim.
- * @return A string view with trailing whitespace removed.
- */
-constexpr std::string_view rtrim(std::string_view s) {
-    const auto not_ws_pos = s.find_last_not_of(" \t\n\r\f\v");
-    return not_ws_pos == std::string_view::npos ? std::string_view() : s.substr(0, not_ws_pos + 1);
-}
-
-/**
- * @brief Return a string view with leading and trailing whitespace removed.
- * @param s A string view to trim.
- * @return A string view with leading and trailing whitespace removed.
- */
-constexpr std::string_view trim(std::string_view s) {
-    return ltrim(rtrim(s));
+    return "[ " + join(v) + " ]";
 }
 
 /**
@@ -181,50 +136,9 @@ constexpr bool ends_with(std::basic_string_view<T> src, std::basic_string_view<T
 }
 /** @} */
 
-template <typename T>
-constexpr T ceil_div(const T& x, const T& y) {
-    return (x == 0 ? 0 : (1 + (x - 1) / y));
-}
+std::string to_lower(std::string_view s);
 
-/**
- * @brief Checks if container contains the specific value.
- *
- * @param container  The container of elements to examine.
- * @param value      Value to compare the elements to.
- * @return True if value found in the container, false otherwise.
- */
-template <typename R, typename V>
-constexpr bool contains(const R& container, const V& value) {
-    return std::find(std::begin(container), std::end(container), value) != std::end(container);
-}
-
-/**
- * @brief multiply vector's values
- * @param vec - vector with values
- * @return result of multiplication
- */
-template <typename Container>
-auto product(const Container& container) {
-    using T = typename Container::value_type;
-    return container.empty() ? T{0} : std::accumulate(container.begin(), container.end(), T{1}, std::multiplies<T>());
-}
-
-/**
- * @brief Removes elements from the container that satisfy the given predicate.
- *
- * @param data      The container from which to remove elements.
- * @param predicate A callable that check element and return true if the element should be removed, or false otherwise.
- */
-template <typename Container, typename PredicateT>
-inline void erase_if(Container& data, const PredicateT& predicate) {
-    for (auto it = std::begin(data); it != std::end(data);) {
-        if (predicate(*it)) {
-            it = data.erase(it);
-        } else {
-            ++it;
-        }
-    }
-}
+std::string to_upper(std::string_view s);
 
 /**
  * @brief Filters lines in a string view based on a given prefix.
@@ -237,12 +151,6 @@ inline void erase_if(Container& data, const PredicateT& predicate) {
  * @return A string containing only the lines that start with the given prefix.
  */
 std::string filter_lines_by_prefix(std::string_view sv, std::string_view prefix);
-
-template <class T = void, class... Args>
-constexpr std::array<std::conditional_t<std::is_void_v<T>, std::common_type_t<Args...>, T>, sizeof...(Args)> make_array(
-    Args&&... args) {
-    return {std::forward<Args>(args)...};
-}
 
 /**
  * @brief A custom stream buffer that provides read-only access to a string view.
@@ -293,75 +201,6 @@ protected:
         return seekoff(pos, std::ios_base::beg, which);
     }
 };
-
-/**
- * @brief Adds two integral values
- *
- * The result value is not valid if overflow detected.
- *
- * @param T       Type of values to add. Must be an integral type.
- * @param x       First value to add.
- * @param y       Second value to add.
- * @param result  Reference to store result value.
- * @return True if overflow occurs, false otherwise
- */
-template <class T>
-constexpr bool add_overflow(T x, T y, T& result) {
-    static_assert(std::is_integral_v<T>, "T must be an integral type");
-#if defined(__GNUC__) || defined(__clang__)
-    return __builtin_add_overflow(x, y, &result);
-#else
-    constexpr auto max = std::numeric_limits<T>::max();
-
-    if constexpr (std::is_unsigned_v<T>) {
-        if (x > max - y) {
-            return true;
-        }
-    } else {
-        constexpr auto min = std::numeric_limits<T>::lowest();
-        if ((y > 0 && x > max - y) || (y < 0 && x < min - y)) {
-            return true;
-        }
-    }
-    result = x + y;
-    return false;
-#endif
-}
-
-/**
- * @brief Multiplies two integral values
- *
- * The result value is not valid if overflow detected.
- *
- * @param T       Type of values to multiply. Must be an integral type.
- * @param x       First value to multiply.
- * @param y       Second value to multiply.
- * @param result  Reference to store result value.
- * @return True if overflow occurs, false otherwise
- */
-template <class T>
-constexpr bool mul_overflow(T x, T y, T& result) {
-    static_assert(std::is_integral_v<T>, "T must be an integral type");
-#if defined(__GNUC__) || defined(__clang__)
-    return __builtin_mul_overflow(x, y, &result);
-#else
-    constexpr auto max = std::numeric_limits<T>::max();
-
-    if constexpr (std::is_unsigned_v<T>) {
-        if (y > 0 && x > max / y) {
-            return true;
-        }
-    } else {
-        constexpr auto min = std::numeric_limits<T>::lowest();
-        if ((x > 0 && y > 0 && x > max / y) || (x > 0 && y < 0 && y < min / x) || (x < 0 && y > 0 && x < min / y) ||
-            (x < 0 && y < 0 && x < max / y)) {
-            return true;
-        }
-    }
-    result = x * y;
-    return false;
-#endif
-}
 
 /**
  * @brief This function attempts to parse the input string view `sv` into a number of type `T`.
