+++ src/common/util/src/os/lin/lin_mmap_object.cpp
@@ -13,8 +13,8 @@
 #include <thread>
 #include <tuple>
 
-#include "openvino/util/common_util.hpp"
 #include "openvino/util/file_util.hpp"
+#include "openvino/util/hash_util.hpp"
 #include "openvino/util/memory.hpp"
 #include "openvino/util/mmap_object.hpp"
